Here we go again, rechange all and fix error loding data on agent page
