+++ client/src/components/AgentComponents/AgentPatientSearchBar.tsx/AgentPatientSearchBar.type.tsx
@@ -1,0 +1,4 @@
+export default interface AgentPatientSearchBarProps {
+  handlePatientSelected: (patientId: number) => void;
+  restriction?: boolean;
+}
