+++ client/src/components/AgentComponents/AgentSearchBar.tsx/AgentSearchBar.type.tsx
@@ -1,0 +1,4 @@
+export default interface AgentSearchBarProps {
+  handleChange: (value: string) => void;
+  inputType?: 'text' | 'number';
+}
